+++ opennn/growing_inputs.cpp
@@ -883,7 +883,7 @@ void GrowingInputs::from_XML(const tinyxml2::XMLDocument& document)
 
         if(element)
         {
-            const double new_maximum_iterations_number = atoi(element->GetText());
+            const size_t new_maximum_iterations_number = atoi(element->GetText());
 
             try
             {
@@ -978,7 +978,7 @@ void GrowingInputs::from_XML(const tinyxml2::XMLDocument& document)
 
         if(element)
         {
-            const double new_maximum_generalization_failures = atof(element->GetText());
+            const size_t new_maximum_generalization_failures = atoi(element->GetText());
 
             try
             {
